1_models_langchain/requirements.txt
langchain[openai]
langchain[anthropic]
langchain[google-genai]
langchain[huggingface]
openai
google-generativeai
transformers
huggingface-hub
python-dotenv
torch
langchain_groq 

